DDCYLS-170 - All integration tests live in the "it" directory under test - build.sbt modified in a lot of places
