+++ build.sbt
@@ -1,4 +1,7 @@
-import uk.gov.hmrc.DefaultBuildSettings.integrationTestSettings
+import play.sbt.PlayImport.ws
+import sbt.Defaults.testTasks
+import uk.gov.hmrc.DefaultBuildSettings.addTestReportOption
+import uk.gov.hmrc.ForkedJvmPerTestSettings.oneForkedJvmPerTest
 import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin.publishingSettings
 
 val appName = "digital-services-tax-frontend"
@@ -6,6 +9,40 @@ val appName = "digital-services-tax-frontend"
 PlayKeys.playDefaultPort := 8740
 
 addCompilerPlugin("org.typelevel" %% "kind-projector" % "0.10.3")
+scalaVersion := "2.12.11"
+ThisBuild / scalacOptions += "-Ypartial-unification"
+routesImport += "uk.gov.hmrc.digitalservicestax.data._"
+
+lazy val microservice = Project(appName, file("."))
+  .enablePlugins(PlayScala, SbtAutoBuildPlugin, SbtGitVersioning, SbtDistributablesPlugin)
+  .disablePlugins(JUnitXmlReportPlugin) //Required to prevent https://github.com/scalatest/scalatest/issues/1427
+  .settings(
+    majorVersion := 0,
+    libraryDependencies ++= AppDependencies.compile ++ AppDependencies.test ++ Seq(ws),
+    scalacOptions += "-Ypartial-unification"
+  )
+  .settings(publishingSettings: _*)
+  .configs(IntegrationTest)
+  .settings(itSettings)
+  .settings(unitTestSettings)
+  .configs(IntegrationTest)
+  .settings(inConfig(IntegrationTest)(Defaults.testSettings): _*)
+  .settings(
+    resolvers ++= Seq(Resolver.bintrayRepo("hmrc", "releases"), Resolver.jcenterRepo),
+    scalacOptions -= "-Xfatal-warnings", // Fail the compilation if there are any warnings.
+    TwirlKeys.templateImports ++= Seq(
+      "ltbs.uniform.{Input => UfInput, _}",
+      "ltbs.uniform.common.web.{Breadcrumbs => UfBreadcrumbs, _}",
+      "ltbs.uniform.interpreters.playframework._",
+      "uk.gov.hmrc.digitalservicestax.views.html.helpers._",
+      "uk.gov.hmrc.digitalservicestax.views.html.Layout",
+      "uk.gov.hmrc.govukfrontend.views.html.components._",
+      "uk.gov.hmrc.hmrcfrontend.views.html.components._",
+      "uk.gov.hmrc.digitalservicestax.views.AdaptMessages.ufMessagesToPlayMessages",
+      "uk.gov.hmrc.digitalservicestax.data._"
+    )
+  )
+  .settings(scoverageSettings)
 
 lazy val scoverageSettings = {
   import scoverage.ScoverageKeys
@@ -19,47 +56,17 @@ lazy val scoverageSettings = {
   )
 }
 
-scalacOptions in ThisBuild += "-Ypartial-unification"
-
-lazy val microservice = Project(appName, file("."))
-  .enablePlugins(play.sbt.PlayScala, SbtAutoBuildPlugin, SbtGitVersioning, SbtDistributablesPlugin)
-  .disablePlugins(JUnitXmlReportPlugin) //Required to prevent https://github.com/scalatest/scalatest/issues/1427
-  .settings(
-    majorVersion := 0,
-    libraryDependencies ++= AppDependencies.compile ++ AppDependencies.test,
-    scalacOptions += "-Ypartial-unification"
-  )
-  .settings(publishingSettings: _*)
-  .configs(IntegrationTest)
-  .settings(itSettings: _*)
-  .settings(resolvers += Resolver.jcenterRepo)
-  .settings(
-    resolvers += Resolver.jcenterRepo,
-    resolvers += Resolver.bintrayRepo("hmrc", "releases"),
-    scalacOptions -= "-Xfatal-warnings", // Fail the compilation if there are any warnings.
-    TwirlKeys.templateImports ++= Seq(
-      "ltbs.uniform.{Input => UfInput, _}",
-      "ltbs.uniform.common.web.{Breadcrumbs => UfBreadcrumbs, _}",
-      "ltbs.uniform.interpreters.playframework._",
-      "uk.gov.hmrc.digitalservicestax.views.html.helpers._",
-      "uk.gov.hmrc.digitalservicestax.views.html.Layout",
-      "uk.gov.hmrc.govukfrontend.views.html.components._",
-      "uk.gov.hmrc.hmrcfrontend.views.html.components._",
-      "uk.gov.hmrc.digitalservicestax.views.AdaptMessages.ufMessagesToPlayMessages"
-    )
-  ).settings(scoverageSettings)
-
-libraryDependencies ++= Seq(
-  ws
+lazy val unitTestSettings = inConfig(Test)(testTasks) ++ Seq(
+  Test / testOptions := Seq(Tests.Filter(name => name startsWith "unit")),
+  Test / unmanagedSourceDirectories := Seq((baseDirectory in Test).value / "test"),
+  Test / parallelExecution := true,
+  Test / fork := true,
+  addTestReportOption(Test, "unit-test-reports")
 )
 
-routesImport += "uk.gov.hmrc.digitalservicestax.data._"
-TwirlKeys.templateImports += "uk.gov.hmrc.digitalservicestax.data._"
-
-scalaVersion := "2.12.11"
-
-val itSettings = Defaults.itSettings ++ Seq(
-  unmanagedSourceDirectories in IntegrationTest += baseDirectory(_ / "it").value,
-  parallelExecution := false,
-  fork := true
+lazy val itSettings = inConfig(IntegrationTest)(testTasks) ++ Seq(
+  IntegrationTest / testOptions := Seq(Tests.Filter(name => name startsWith "it")),
+  IntegrationTest / parallelExecution := false,
+  IntegrationTest / fork := false,
+  addTestReportOption(IntegrationTest, "integration-test-reports")
 )
