+++ test/unit/uk/gov/hmrc/digitalservicestaxfrontend/util/TestWiring.scala
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2021 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package unit.uk.gov.hmrc.digitalservicestaxfrontend.util
+
+import org.scalatestplus.mockito.MockitoSugar
+import play.api.{Configuration, Environment}
+
+import java.io.File
+import java.time.Clock
+
+trait TestWiring extends MockitoSugar {
+  val appName: String = configuration.get[String]("appName")
+
+  lazy val configuration: Configuration = Configuration.load(environment, Map(
+    "auditing.enabled" -> "false",
+    "services.auth.port" -> "11111"
+  ))
+  lazy val environment: Environment = Environment.simple(new File("."))
+  lazy val mode = environment.mode
+
+  implicit def clock: Clock = Clock.systemDefaultZone()
+}
